Fix bugs related to Timer
than use #startPauseCallback directly as this was causing the
input readonly toggle to fail.

Saved the edit time logic for later since it was taking longer
than expected.

The time element was updated to use 3 inputs to make the time
editable. This was causing 2 Timer#constructor specs to fail
because they were checking for the innerText of the time element
which no longer holds the time string. This was fixed.