Finalize background scroll logic
I created the scroll logic through events. A scroll event is
dispatched when the Background#scroll is called with a direction.
The direction is saved on the custom scroll event for use in the
listener.

The scroll listener transforms the direction into a direction
delta that is used to increment the position delta on the
background object at each interval.