Fix background scroll bugs
I changed the listener that increments the position of a background
on an interval to be set only once. This listener was being
triggered each time the scroll method was called on a Background
object. Because each object would add the listener on creation,
and because each scroll call would trigger all listeners,
the scrolling functionality was not functioning properly. The
scrolling interval was being set twice, once for each added
listener (I am using 2 background objects). Then, the first
interval timeout ID was lost and so it was not being cleared.

The same issue was occuring with when clearing the timeouts. It
wasn't a problem calling this logic multiple times as it did not
affect the expected behavior. However, because multiple intervals
were being set per background object, and because only one timeout
ID was being held per background object, not all intervals were
being cleared.