Refactor listeners and Game intialization
I refactored the listeners for objects that are constructed more
than once. The listeners listen for a custom event, and in the
event they catch, is the instance of the object that triggered
the event.

I also add all listeners after constructing the game object so
that the listeners aren't added more than once.

Within the game constructor, I moved settup logic to an init
function, and broke down the startup code more from there.