Scale background on resize
To scale the size of the backgrounds on resize, I draw the images
over the entire canvas and scale their positions.

Position scaling is done using the initial width and height of the
canvas on start up. I create a ratio by dividing the current
canvas dimensions with the initial dimensions. This ratio helps
calculate the position of x and y for a background on the canvas.
This ratio is required because the saved position of a background
is always calculated relative to the initial canvas size.

Before, without adjusting the position with the ratio, the
positions were relative to the initial size, so scrolling would
restart based on different dimensions. With the use of ratios
to scale positions, scrolling restarts at the correct positions.

Lastly, to restart the scroll at the correct position, the
position deltas are cycled within the range of the initial
dimensions. This is because the position deltas are scaled
up/down from the initial dimensions and so they reach the end
of their cycle at the end of the range of the initial
dimensions.