Refactor Timer, Dino, Background, Drawable and Game
Timer has been refactored to not use listeners. Game was refactored
to use Timer's built in functions.

The Game's listeners call the Game functions that delegate the work
to be done to Timer and Dino.

The Drawable class and all descendants now have a reference to the
canvas element and its context rather than the Canvas object. This
was done to reduce dependencies and only use the required data
within Drawable objects.