Create DinoMovable base functionality
The Movable class handles the basic x and y repositioning based
on the given direction (1 or -1). It also uses a special function
called addLivePropsFrom to add the properties of the given
drawable to the current instance. This is done by taking each
property of the drawable and creating a function with the same
name on the current instance that simply returns the corresponding
drawable property. This is necessary to avoid inherting from
drawable in order to access its properties. It's also necessary
to avoid using 'this.drawable' everywhere and polluting the code
with train wrecks.

The DinoMovable class extends the Movable class and handles the
special DinoSprite repositioning logic.